add calculate sale percentage in retail

+++ src/main/java/com/elleined/marketplaceapi/controller/seller/WholeSaleSellerController.java
@@ -56,7 +56,7 @@ public class WholeSaleSellerController {
     @PatchMapping("/{productId}/sale")
     public WholeSaleProductDTO saleProduct(@PathVariable("currentUserId") int sellerId,
                                            @PathVariable("productId") int productId,
-                                           @RequestParam("salePrice") BigDecimal salePrice ) {
+                                           @RequestParam("salePrice") double salePrice ) {
 
         User seller = userService.getById(sellerId);
         WholeSaleProduct wholeSaleProduct = wholeSaleProductService.getById(productId);
